extract code from WASM lib.rs into sub-modules;
