+++ bindings/ergo-wallet-lib-wasm/src/address.rs
@@ -1,0 +1,67 @@
+//! Address types
+
+use sigma_tree::chain;
+use wasm_bindgen::prelude::*;
+
+/**
+ * An address is a short string corresponding to some script used to protect a box. Unlike (string-encoded) binary
+ * representation of a script, an address has some useful characteristics:
+ *
+ * - Integrity of an address could be checked., as it is incorporating a checksum.
+ * - A prefix of address is showing network and an address type.
+ * - An address is using an encoding (namely, Base58) which is avoiding similarly l0Oking characters, friendly to
+ * double-clicking and line-breaking in emails.
+ *
+ *
+ *
+ * An address is encoding network type, address type, checksum, and enough information to watch for a particular scripts.
+ *
+ * Possible network types are:
+ * Mainnet - 0x00
+ * Testnet - 0x10
+ *
+ * For an address type, we form content bytes as follows:
+ *
+ * P2PK - serialized (compressed) public key
+ * P2SH - first 192 bits of the Blake2b256 hash of serialized script bytes
+ * P2S  - serialized script
+ *
+ * Address examples for testnet:
+ *
+ * 3   - P2PK (3WvsT2Gm4EpsM9Pg18PdY6XyhNNMqXDsvJTbbf6ihLvAmSb7u5RN)
+ * ?   - P2SH (rbcrmKEYduUvADj9Ts3dSVSG27h54pgrq5fPuwB)
+ * ?   - P2S (Ms7smJwLGbUAjuWQ)
+ *
+ * for mainnet:
+ *
+ * 9  - P2PK (9fRAWhdxEsTcdb8PhGNrZfwqa65zfkuYHAMmkQLcic1gdLSV5vA)
+ * ?  - P2SH (8UApt8czfFVuTgQmMwtsRBZ4nfWquNiSwCWUjMg)
+ * ?  - P2S (4MQyML64GnzMxZgm, BxKBaHkvrTvLZrDcZjcsxsF7aSsrN73ijeFZXtbj4CXZHHcvBtqSxQ)
+ *
+ *
+ * Prefix byte = network type + address type
+ *
+ * checksum = blake2b256(prefix byte ++ content bytes)
+ *
+ * address = prefix byte ++ content bytes ++ checksum
+ *
+ */
+#[wasm_bindgen]
+pub struct Address(chain::address::Address);
+
+#[wasm_bindgen]
+impl Address {
+    /// Decode (base58) testnet address from string
+    pub fn from_testnet_str(s: &str) -> Result<Address, JsValue> {
+        chain::address::AddressEncoder::new(chain::address::NetworkPrefix::Testnet)
+            .parse_address_from_str(s)
+            .map(Address)
+            .map_err(|e| JsValue::from_str(&format!("{}", e)))
+    }
+}
+
+impl Into<chain::address::Address> for Address {
+    fn into(self) -> chain::address::Address {
+        self.0
+    }
+}
