+++ bindings/ergo-wallet-lib-wasm/src/contract.rs
@@ -1,0 +1,24 @@
+use sigma_tree::chain;
+use wasm_bindgen::prelude::*;
+
+use crate::address::Address;
+
+/// Defines the contract(script) that will be guarding box contents
+#[wasm_bindgen]
+pub struct Contract(chain::contract::Contract);
+
+#[wasm_bindgen]
+impl Contract {
+    /// create new contract that allow spending of the guarded box by a given recipient ([`Address`])
+    pub fn pay_to_address(recipient: Address) -> Result<Contract, JsValue> {
+        chain::contract::Contract::pay_to_address(recipient.into())
+            .map_err(|e| JsValue::from_str(&format!("{}", e)))
+            .map(Contract)
+    }
+}
+
+impl Into<chain::contract::Contract> for Contract {
+    fn into(self) -> chain::contract::Contract {
+        self.0
+    }
+}
