+++ bindings/ergo-wallet-lib-wasm/src/ergo_state_ctx.rs
@@ -1,0 +1,14 @@
+use sigma_tree::chain;
+use wasm_bindgen::prelude::*;
+
+/// TBD
+#[wasm_bindgen]
+pub struct ErgoStateContext(chain::ergo_state_context::ErgoStateContext);
+
+#[wasm_bindgen]
+impl ErgoStateContext {
+    /// empty (dummy) context (for signing P2PK tx only)
+    pub fn dummy() -> ErgoStateContext {
+        ErgoStateContext(chain::ergo_state_context::ErgoStateContext::dummy())
+    }
+}
