+++ bindings/ergo-wallet-lib-wasm/src/lib.rs
@@ -12,283 +12,13 @@
 // #![deny(missing_docs)]
 #![allow(unused_variables)]
 
-use std::convert::TryFrom;
-
-use sigma_tree::chain;
-use sigma_tree::wallet;
-
-mod utils;
-
-use wasm_bindgen::prelude::*;
-
-/**
- * An address is a short string corresponding to some script used to protect a box. Unlike (string-encoded) binary
- * representation of a script, an address has some useful characteristics:
- *
- * - Integrity of an address could be checked., as it is incorporating a checksum.
- * - A prefix of address is showing network and an address type.
- * - An address is using an encoding (namely, Base58) which is avoiding similarly l0Oking characters, friendly to
- * double-clicking and line-breaking in emails.
- *
- *
- *
- * An address is encoding network type, address type, checksum, and enough information to watch for a particular scripts.
- *
- * Possible network types are:
- * Mainnet - 0x00
- * Testnet - 0x10
- *
- * For an address type, we form content bytes as follows:
- *
- * P2PK - serialized (compressed) public key
- * P2SH - first 192 bits of the Blake2b256 hash of serialized script bytes
- * P2S  - serialized script
- *
- * Address examples for testnet:
- *
- * 3   - P2PK (3WvsT2Gm4EpsM9Pg18PdY6XyhNNMqXDsvJTbbf6ihLvAmSb7u5RN)
- * ?   - P2SH (rbcrmKEYduUvADj9Ts3dSVSG27h54pgrq5fPuwB)
- * ?   - P2S (Ms7smJwLGbUAjuWQ)
- *
- * for mainnet:
- *
- * 9  - P2PK (9fRAWhdxEsTcdb8PhGNrZfwqa65zfkuYHAMmkQLcic1gdLSV5vA)
- * ?  - P2SH (8UApt8czfFVuTgQmMwtsRBZ4nfWquNiSwCWUjMg)
- * ?  - P2S (4MQyML64GnzMxZgm, BxKBaHkvrTvLZrDcZjcsxsF7aSsrN73ijeFZXtbj4CXZHHcvBtqSxQ)
- *
- *
- * Prefix byte = network type + address type
- *
- * checksum = blake2b256(prefix byte ++ content bytes)
- *
- * address = prefix byte ++ content bytes ++ checksum
- *
- */
-#[wasm_bindgen]
-pub struct Address(chain::address::Address);
-
-#[wasm_bindgen]
-impl Address {
-    /// Decode (base58) testnet address from string
-    pub fn from_testnet_str(s: &str) -> Result<Address, JsValue> {
-        chain::address::AddressEncoder::new(chain::address::NetworkPrefix::Testnet)
-            .parse_address_from_str(s)
-            .map(Address)
-            .map_err(|e| JsValue::from_str(&format!("{}", e)))
-    }
-}
-
-/// Secret key for the prover
-#[wasm_bindgen]
-#[derive(PartialEq, Debug, Clone)]
-pub struct SecretKey(wallet::secret_key::SecretKey);
-
-#[wasm_bindgen]
-impl SecretKey {
-    /// generate random key
-    pub fn random_dlog() -> SecretKey {
-        SecretKey(wallet::secret_key::SecretKey::random_dlog())
-    }
-}
-
-/// Collection of ErgoBox'es
-#[wasm_bindgen]
-pub struct ErgoBoxes(Vec<chain::ergo_box::ErgoBox>);
-
-#[wasm_bindgen]
-impl ErgoBoxes {
-    /// parse ErgoBox array from json
-    #[allow(clippy::boxed_local)]
-    pub fn from_boxes(_boxes: Box<[JsValue]>) -> ErgoBoxes {
-        // box in boxes.into_iter() {
-        //     let _box: chain::ErgoBoxCandidate = jbox.into_serde().unwrap();
-        // }
-        ErgoBoxes(vec![])
-    }
-}
-
-/// Collection of ErgoBoxCandidates
-#[wasm_bindgen]
-pub struct ErgoBoxCandidates(Vec<chain::ergo_box::ErgoBoxCandidate>);
-
-#[wasm_bindgen]
-impl ErgoBoxCandidates {
-    /// Create new outputs
-    #[wasm_bindgen(constructor)]
-    pub fn new(box_candidate: ErgoBoxCandidate) -> ErgoBoxCandidates {
-        ErgoBoxCandidates(vec![box_candidate.0])
-    }
-}
-
-/// Box (aka coin, or an unspent output) is a basic concept of a UTXO-based cryptocurrency.
-/// In Bitcoin, such an object is associated with some monetary value (arbitrary,
-/// but with predefined precision, so we use integer arithmetic to work with the value),
-/// and also a guarding script (aka proposition) to protect the box from unauthorized opening.
-///
-/// In other way, a box is a state element locked by some proposition (ErgoTree).
-///
-/// In Ergo, box is just a collection of registers, some with mandatory types and semantics,
-/// others could be used by applications in any way.
-/// We add additional fields in addition to amount and proposition~(which stored in the registers R0 and R1).
-/// Namely, register R2 contains additional tokens (a sequence of pairs (token identifier, value)).
-/// Register R3 contains height when block got included into the blockchain and also transaction
-/// identifier and box index in the transaction outputs.
-/// Registers R4-R9 are free for arbitrary usage.
-///
-/// A transaction is unsealing a box. As a box can not be open twice, any further valid transaction
-/// can not be linked to the same box.
-#[wasm_bindgen]
-pub struct ErgoBoxCandidate(chain::ergo_box::ErgoBoxCandidate);
-
-#[wasm_bindgen]
-impl ErgoBoxCandidate {
-    /// make a new box with:
-    /// `value` - amount of money associated with the box
-    /// `contract` - guarding contract([`Contract`]), which should be evaluated to true in order
-    /// to open(spend) this box
-    /// `creation_height` - height when a transaction containing the box is created.
-    /// It should not exceed height of the block, containing the transaction with this box.
-    #[wasm_bindgen(constructor)]
-    pub fn new(value: BoxValue, creation_height: u32, contract: Contract) -> ErgoBoxCandidate {
-        let ergo_tree = contract.0.get_ergo_tree();
-        let b = chain::ergo_box::ErgoBoxCandidate::new(value.0, ergo_tree, creation_height);
-        ErgoBoxCandidate(b)
-    }
-
-    // JSON representation
-    // pub fn to_json(&self) -> Result<JsValue, JsValue> {
-    //     JsValue::from_serde(&self.0).map_err(|e| JsValue::from_str(&format!("{}", e)))
-    // }
-}
-
-/// Defines the contract(script) that will be guarding box contents
-#[wasm_bindgen]
-pub struct Contract(chain::contract::Contract);
-
-#[wasm_bindgen]
-impl Contract {
-    /// create new contract that allow spending of the guarded box by a given recipient ([`Address`])
-    pub fn pay_to_address(recipient: Address) -> Result<Contract, JsValue> {
-        chain::contract::Contract::pay_to_address(recipient.0)
-            .map_err(|e| JsValue::from_str(&format!("{}", e)))
-            .map(Contract)
-    }
-}
-
-/**
- * ErgoTransaction is an atomic state transition operation. It destroys Boxes from the state
- * and creates new ones. If transaction is spending boxes protected by some non-trivial scripts,
- * its inputs should also contain proof of spending correctness - context extension (user-defined
- * key-value map) and data inputs (links to existing boxes in the state) that may be used during
- * script reduction to crypto, signatures that satisfies the remaining cryptographic protection
- * of the script.
- * Transactions are not encrypted, so it is possible to browse and view every transaction ever
- * collected into a block.
- */
-#[wasm_bindgen]
-pub struct Transaction(chain::transaction::Transaction);
-
-#[wasm_bindgen]
-impl Transaction {
-    /// JSON representation
-    pub fn to_json(&self) -> Result<JsValue, JsValue> {
-        JsValue::from_serde(&self.0).map_err(|e| JsValue::from_str(&format!("{}", e)))
-    }
-}
-
-/// TBD
-#[wasm_bindgen]
-pub struct ErgoStateContext(chain::ergo_state_context::ErgoStateContext);
-
-#[wasm_bindgen]
-impl ErgoStateContext {
-    /// empty (dummy) context (for signing P2PK tx only)
-    pub fn dummy() -> ErgoStateContext {
-        ErgoStateContext(chain::ergo_state_context::ErgoStateContext::dummy())
-    }
-}
-
-/// TBD
-#[wasm_bindgen]
-pub struct Wallet();
-
-#[wasm_bindgen]
-impl Wallet {
-    /// Create wallet instance loading secret key from mnemonic
-    pub fn from_mnemonic(_mnemonic_phrase: &str, _mnemonic_pass: &str) -> Wallet {
-        Wallet()
-    }
-
-    /// Sign a transaction:
-    /// `boxes_to_spend` - unspent boxes [`ErgoBoxCandidate`] used as inputs in the transaction
-    #[wasm_bindgen]
-    pub fn sign_transaction(
-        &self,
-        _state_context: ErgoStateContext,
-        tx: UnsignedTransaction,
-        boxes_to_spend: ErgoBoxes,
-        data_boxes: ErgoBoxes,
-    ) -> Result<Transaction, JsValue> {
-        // not implemented, see https://github.com/ergoplatform/sigma-rust/issues/34
-        Err(JsValue::from_str("Not yet implemented"))
-    }
-}
-
-#[wasm_bindgen]
-pub struct BoxValue(chain::ergo_box::box_value::BoxValue);
-
-#[wasm_bindgen]
-impl BoxValue {
-    #[wasm_bindgen]
-    pub fn from_u32(v: u32) -> Result<BoxValue, JsValue> {
-        Ok(BoxValue(
-            chain::ergo_box::box_value::BoxValue::try_from(v as u64)
-                .map_err(|e| JsValue::from_str(&format!("{}", e)))?,
-        ))
-    }
-}
-
-#[wasm_bindgen]
-pub struct UnsignedTransaction(chain::transaction::unsigned::UnsignedTransaction);
-
-#[wasm_bindgen]
-impl UnsignedTransaction {
-    #[wasm_bindgen]
-    pub fn dummy() -> UnsignedTransaction {
-        UnsignedTransaction(chain::transaction::unsigned::UnsignedTransaction::new(
-            vec![],
-            vec![],
-            vec![],
-        ))
-    }
-}
-
-#[wasm_bindgen]
-pub struct TxBuilder(wallet::tx_builder::TxBuilder);
-
-#[wasm_bindgen]
-impl TxBuilder {
-    #[wasm_bindgen]
-    pub fn new(
-        inputs: ErgoBoxes,
-        output_candidates: ErgoBoxCandidates,
-        current_height: u32,
-        fee_amount: BoxValue,
-    ) -> Result<TxBuilder, JsValue> {
-        Err(JsValue::from_str("Not yet implemented"))
-    }
-
-    #[wasm_bindgen]
-    pub fn with_change_sent_to(
-        &self,
-        change_address: &Address,
-        min_change_value: BoxValue,
-    ) -> TxBuilder {
-        todo!()
-    }
-
-    #[wasm_bindgen]
-    pub fn build(&self) -> Result<UnsignedTransaction, JsValue> {
-        Err(JsValue::from_str("Not yet implemented"))
-    }
-}
+pub mod address;
+pub mod box_coll;
+pub mod contract;
+pub mod ergo_box;
+pub mod ergo_state_ctx;
+pub mod secret_key;
+pub mod transaction;
+pub mod tx_builder;
+pub mod utils;
+pub mod wallet;
